requirements.txt
aiogram>=3.17,<4
chromadb>=0.5,<1
openai>=1.60,<2
anthropic>=0.42,<1
python-dotenv>=1.0,<2
asyncpg>=0.30,<1
sqlalchemy[asyncio]>=2.0,<3
python-docx>=1.1,<2
openpyxl>=3.1,<4
reportlab>=4.2,<5
pydantic>=2.10,<3
pydantic-settings>=2.7,<3
Pillow>=10.0,<12
pdfplumber>=0.11,<1
